requirements.txt
# HAKA Voice Assistant v4 — Python Dependencies
# Install with: pip install -r requirements.txt

# Core API clients
groq>=0.8.0
ollama>=0.2.0

# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# Audio
sounddevice>=0.4.6
numpy>=1.24.0
SpeechRecognition>=3.10.0

# TTS engines
pyttsx3>=2.90
gTTS>=2.5.0
edge-tts>=6.1.9

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# Optional: WebRTC VAD for better silence detection
# webrtcvad>=2.0.10   # Uncomment if you have a C compiler
